[refactor] #100 Matching 도메인 리팩토링
